+++ R/accessors-second.r
@@ -8,6 +8,7 @@ NULL
 #'
 #' @export
 #' @param x a date-time object
+#' @param value numeric value to be assigned
 #' @return the seconds element of x as a decimal number
 #' @keywords utilities manip chron methods
 #' @examples
@@ -31,17 +32,37 @@ second.Period <- function(x) {
 }
 
 #' @rdname second
-#' @param value numeric value to be assigned
 #' @export
-"second<-" <- function(x, value) {
+setGeneric("second<-",
+  function (x, value) standardGeneric("second<-"),
+  useAsDefault = function(x, value) {
+    y <- update_date_time(as.POSIXct(x), seconds = value)
+    reclass_date(y, x)
+  }
+)
+
+#' @export
+setMethod("second<-", "Duration", function(x, value) {
   x <- x + seconds(value - second(x))
-}
-
-
-setGeneric("second<-")
+})
 
 #' @export
 setMethod("second<-", signature("Period"), function(x, value) {
   slot(x, ".Data") <- value
   x
+})
+
+#' @export
+setMethod("second<-", "Interval", function(x, value) {
+  x <- x + seconds(value - second(x))
+})
+
+#' @export
+setMethod("second<-", "POSIXt", function(x, value) {
+  update.POSIXt(x, seconds = value)
+})
+
+#' @export
+setMethod("second<-", "Date", function(x, value) {
+  update.Date(x, seconds = value)
 })